+++ library/src/main/java/com/getstream/sdk/chat/adapter/MediaAttachmentSelectedAdapter.java
@@ -12,6 +12,7 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
 import com.getstream.sdk.chat.R;
+import com.getstream.sdk.chat.StreamChat;
 import com.getstream.sdk.chat.databinding.StreamItemAttachedMediaBinding;
 import com.getstream.sdk.chat.model.Attachment;
 import com.getstream.sdk.chat.model.ModelType;
@@ -79,12 +80,12 @@ public class MediaAttachmentSelectedAdapter extends RecyclerView.Adapter<MediaAt
                 if (file.exists()) {
                     Uri imageUri = Uri.fromFile(file);
                     Glide.with(context)
-                            .load(imageUri)
+                            .load(StreamChat.getInstance(context).getUploadStorage().signGlideUrl(imageUri.toString()))
                             .into(binding.ivMedia);
                 }
             } else if (!TextUtils.isEmpty(attachment.getImageURL())) {
                 Glide.with(context)
-                        .load(attachment.getImageURL())
+                        .load(StreamChat.getInstance(context).getUploadStorage().signGlideUrl(attachment.getImageURL()))
                         .into(binding.ivMedia);
             } else {
                 try {